Gb 109 img upload, 비즈니스 로직 분리 (#26)
* [GB-109] feat : img 업로드 및 비즈니스 로직 분리

* [GB-109] fix : baseURL 수정
